requirements.txt
asgiref==3.9.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
djangorestframework==3.16.1
filelock==3.18.0
gunicorn==23.0.0
packaging==25.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.33.1
whitenoise==6.9.0
